final commit
I finished it. I need Clawson help but I figured out that I had the wrong syntax on the onchange and I thought it was right because it was showing numbers but If i had looked at what it was supposed to look like it wasn't supposed to be showing numbers yet.